Migrated all functions from `dfda.ts` to `dfdaActions.ts` to streamline dependencies. Updated all corresponding file imports to use the new module path.
Took 20 minutes
