+++ app/api/dfda/passwordless-login/route.ts
@@ -1,7 +1,5 @@
 import { z } from "zod"
 
-import { getUserIdServer } from "@/lib/api/getUserIdServer"
-import { dfdaPOST } from "@/lib/dfda"
 import { handleError } from "@/lib/errorHandler"
 
 const requestBodySchema = z.object({
