+++ app/dfda/safe/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation"
 
-import { getSafeRedirectUrl } from "@/lib/dfda"
+import { getSafeRedirectUrl } from "../dfdaActions"
 import { getCurrentUser } from "@/lib/session"
 import { LoginPromptButton } from "@/components/LoginPromptButton"
